REL-245 - use DomainObject RelayCount.

+++ relayD-web/src/main/java/com/relayd/web/pagebean/event/RelayEventEditPageBean.java
@@ -126,7 +126,8 @@ public class RelayEventEditPageBean implements Serializable {
 	}
 
 	public Integer getNumberOfRelays() {
-		return workingRelayEvent.getMaxNumberOfRelays();
+		// TODO
+		return workingRelayEvent.getMaxNumberOfRelays().intValue();
 	}
 
 	void persistRelayEvent() {
